Website Progress!
I've done quite a bit of learning and application today! I learned from freecodecamp.org how to modify containers pretty well, and used that knowledge as well as other online resources to start making the website look how I want it to. I added 5 different sets of containers to my website (2 titles, one for some plaintext, one for the grid, and a group for the items in the grid) and used a grid to try and make changing the size of the browser window not destroy the flow of the site.