Reformatting, adding pictures, changing color palette.
I added 2 pictures, one from the west side of campus and one from the east side of campus for the area selector buttons. I made it so that when any part of said selectors are hovered the text within them changes colors, as well as made it so that the whole selector image is a link, rather than just the text box within it. Finally, I changed the color palette of the front page to make it look better.