Comments and Removal of Unnecessary
I added comments to each section of the stylesheet for the front page in order for it to be understood easily, reformatted the actual contents of the formatting sheet so every rule is ordered, and removed many unnecessary rules and elements from the html and css documents for the front page! Next up: the actual pages where the photography will be held or reformatting the feedback page.